requirements.txt
# Dependencies will be added later
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
requests
